src/requirements.txt
boto3
ish_parser==0.0.24
pandas==1.4.0
prophet==1.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
requests
scipy==1.8.0
SQLAlchemy==1.4.39
